make leftmenu code more readable
