+++ pytripgui/leftmenu.py
@@ -49,38 +49,39 @@ class LeftMenuTree(wx.TreeCtrl):
                              "plans": [{"text": "New plan with ROIs",
                                         "callback": self.new_plan},
                                        {"text": "New empty plan",
-                                        "callback": self.new_empty_plan}],
+                                        "callback": self.new_empty_plan},
+                                       {"text": "New plan from .exec",
+                                        "callback": self.new_plan_from_exec}],
                              "TripPlan": [{"text": "Set Active",
                                            "callback": self.plan_set_active},
                                           {"text": "Add Field",
                                            "callback": self.plan_add_field},
                                           {"text": "Export",
                                            "type": "submenu",
-                                           "submenu": [
-                                               {"text": "Plan data",
-                                                "callback": self.plan_export_exec},
-                                               {"text": "ROI cube",
-                                                "callback": self.plan_export_cube}
-                                           ]}, {"text": "Import",
-                                                "type": "submenu",
-                                                "submenu": [
-                                                    {"text": "Dose (.dos)",
-                                                     "callback": self.plan_load_dose_voxelplan},
-                                                    {"text": "LET (.dosemlet.dos)",
-                                                     "callback": self.plan_load_let_voxelplan}
-                                                ]}, {"text": "Calculate",
-                                                     "type": "submenu",
-                                                     "submenu": [
-                                                         {"text": "Execute TRiP",
-                                                          "callback": self.plan_run_trip}
-                                                     ]}, {"text": "Edit",
-                                                          "callback": self.edit_label}, {"text": "Delete",
-                                                                                         "callback": self.delete_plan},
+                                           "submenu": [{"text": "Plan data",
+                                                        "callback": self.plan_export_exec},
+                                                       {"text": "ROI cube",
+                                                        "callback": self.plan_export_cube}]},
+                                          {"text": "Import",
+                                           "type": "submenu",
+                                           "submenu": [{"text": "Dose (.dos)",
+                                                        "callback": self.plan_load_dose_voxelplan},
+                                                       {"text": "LET (.dosemlet.dos)",
+                                                        "callback": self.plan_load_let_voxelplan}]},
+                                          {"text": "Calculate",
+                                           "type": "submenu",
+                                           "submenu": [{"text": "Execute TRiP",
+                                                        "callback": self.plan_run_trip}]},
+                                          {"text": "Edit",
+                                           "callback": self.edit_label},
+                                          {"text": "Delete",
+                                           "callback": self.delete_plan},
                                           {"text": "Properites",
                                            "callback": self.tripplan_properties}],
                              "DoseCube": [{"text": "Delete",
-                                           "callback": self.plan_remove_dose}, {"text": "Set Active For Plan",
-                                                                                "callback": self.plan_set_active_dose},
+                                           "callback": self.plan_remove_dose},
+                                          {"text": "Set Active For Plan",
+                                           "callback": self.plan_set_active_dose},
                                           {"text": "Properties",
                                            "callback": self.plan_dose_properties}],
                              "LETCube": [{"text": "Delete",
@@ -89,32 +90,39 @@ class LeftMenuTree(wx.TreeCtrl):
                              "TripVoi": [{"text": "Select",
                                           "type": "check",
                                           "value": "is_selected",
-                                          "callback": self.toggle_selected_voi}, {"text": "Target",
-                                                                                  "type": "check",
-                                                                                  "value": "is_target",
-                                                                                  "callback": self.plan_toggle_target},
+                                          "callback": self.toggle_selected_voi},
+                                         {"text": "Target",
+                                          "type": "check",
+                                          "value": "is_target",
+                                          "callback": self.plan_toggle_target},
                                          {"text": "OAR",
                                           "type": "check",
                                           "value": "is_oar",
-                                          "callback": self.plan_toggle_oar}, {"text": "Move Up",
-                                                                              "callback":
-                                                                              self.plan_up_voi}, {"text": "Move Down",
-                                                                                                  "callback":
-                                                                                                  self.plan_down_voi},
+                                          "callback": self.plan_toggle_oar},
+                                         {"text": "Move Up",
+                                          "callback":
+                                          self.plan_up_voi},
+                                         {"text": "Move Down",
+                                          "callback":
+                                          self.plan_down_voi},
                                          {"text": "Delete",
-                                          "callback": self.plan_delete_voi}, {"text": "Properties",
-                                                                              "callback": self.tripvoi_properties}],
+                                          "callback": self.plan_delete_voi},
+                                         {"text": "Properties",
+                                          "callback": self.tripvoi_properties}],
                              "MainVoi": [{"text": "Select",
                                           "type": "check",
                                           "value": "is_selected",
-                                          "callback": self.toggle_selected_voi}, {"text": "Add To Plan",
-                                                                                  "type": "submenu",
-                                                                                  "submenu": self.plan_submenu}],
+                                          "callback": self.toggle_selected_voi},
+                                         {"text": "Add To Plan",
+                                          "type": "submenu",
+                                          "submenu": self.plan_submenu}],
                              "FieldCollection": [{"text": "Add Field",
                                                   "callback": self.plan_add_field}],
                              "Field": [{"text": "Delete",
-                                        "callback": self.plan_delete_field}, {"text": "Properties",
-                                                                              "callback": self.plan_properties_field}]}
+                                        "callback": self.plan_delete_field},
+                                       {"text": "Properties",
+                                        "callback": self.plan_properties_field}]}
+
         pub.subscribe(self.on_patient_load, "patient.load")
         pub.subscribe(self.voi_added, "patient.voi.added")
         pub.subscribe(self.plan_added, "plan.new")
