+++ pytripgui/tripexecparser_old.py
